Rename autocomplete to search_box
Autocomplete is now optional and the same partial can be used to
generate a simpler search_box, therefore the rename. The partial is
one only and the shape is the same. If there is a `data` parameter, it
means the search_box implements a remote autocompletion, otherwise it is
just an input type='search' box. Using the same partial, the UI/UX is consistent.